backend/templates_app/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import EmailTemplateViewSet

router = DefaultRouter()
router.register(r'', EmailTemplateViewSet, basename='email-template')

urlpatterns = [
    path('', include(router.urls)),
]
