backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt[crypto]>=5.3
django-cors-headers>=4.3
django-filter>=24.3
psycopg2-binary>=2.9
celery[redis]>=5.4
redis>=5.0
boto3>=1.35
cryptography>=43.0
python-decouple>=3.8
gunicorn>=22.0
